Adds globe display functionality using a priority queue.
The globe uses a priority queue to display messages. When the queue is empty it
will display a random statistic from the Stats Collection. If there aren't any
new messages the old messages will alternate with the stats one at a time.

If new messages arrive, they will be placed first on the queue. Those messages
will be displayed first and then the previous behaviour will resume.